git-svn-id: svn://svn.debian.org/svn/pkg-samba/branches/samba/upstream@2052 fc4039ab-9d04-0410-8cac-899223bdd6b0

+++ source/script/tests/selftest.sh
@@ -6,7 +6,7 @@ if [ $# != 3 ]; then
 fi
 
 SMBTORTURE4=$3
-TESTS=$2
+SUBTESTS=$2
 
 ##
 ## create the test directory
@@ -27,9 +27,12 @@ export TORTURE_MAXTIME
 ## setup the various environment variables we need
 ##
 
+WORKGROUP=SAMBA-TEST
 SERVER=localhost2
 SERVER_IP=127.0.0.2
 USERNAME=`PATH=/usr/ucb:$PATH whoami`
+USERID=`PATH=/usr/ucb:$PATH id | cut -d ' ' -f1 | sed -e 's/uid=\([0-9]*\).*/\1/g'`
+GROUPID=`PATH=/usr/ucb:$PATH id | cut -d ' ' -f2 | sed -e 's/gid=\([0-9]*\).*/\1/g'`
 PASSWORD=test
 
 SRCDIR="`dirname $0`/../.."
@@ -46,8 +49,12 @@ PRIVATEDIR=$PREFIX_ABS/private
 LOCKDIR=$PREFIX_ABS/lockdir
 LOGDIR=$PREFIX_ABS/logs
 SOCKET_WRAPPER_DIR=$PREFIX/sw
-CONFIGURATION="-s $CONFFILE"
+CONFIGURATION="--configfile $CONFFILE"
 SAMBA4CONFIGURATION="-s $SAMBA4CONFFILE"
+NSS_WRAPPER_PASSWD="$PRIVATEDIR/passwd"
+NSS_WRAPPER_GROUP="$PRIVATEDIR/group"
+WINBINDD_SOCKET_DIR=$PREFIX_ABS/winbindd
+WINBINDD_PRIV_PIPE_DIR=$LOCKDIR/winbindd_privileged
 
 export PREFIX PREFIX_ABS
 export CONFIGURATION CONFFILE SAMBA4CONFIGURATION SAMBA4CONFFILE
@@ -55,11 +62,26 @@ export PATH SOCKET_WRAPPER_DIR DOMAIN
 export PRIVATEDIR LIBDIR PIDDIR LOCKDIR LOGDIR SERVERCONFFILE
 export SRCDIR SCRIPTDIR BINDIR
 export USERNAME PASSWORD
-export SMBTORTURE4
-export SERVER SERVER_IP
+export WORKGROUP SERVER SERVER_IP
+export NSS_WRAPPER_PASSWD NSS_WRAPPER_GROUP
+export WINBINDD_SOCKET_DIR WINBINDD_PRIV_PIPE_DIR
 
 PATH=bin:$PATH
 export PATH
+
+if test x"$LD_LIBRARY_PATH" != x""; then
+	LD_LIBRARY_PATH="$BINDIR:$LD_LIBRARY_PATH"
+else
+	LD_LIBRARY_PATH="$BINDIR"
+fi
+echo "LD_LIBRARY_PATH=$LD_LIBRARY_PATH"
+export LD_LIBRARY_PATH
+
+SAMBA4BINDIR=`dirname $SMBTORTURE4`
+SAMBA4SHAREDDIR="$SAMBA4BINDIR/shared"
+
+export SAMBA4SHAREDDIR
+export SMBTORTURE4
 
 ##
 ## verify that we were built with --enable-socket-wrapper
@@ -73,12 +95,24 @@ if test "x`smbd -b | grep SOCKET_WRAPPER`" = "x"; then
 	exit 1
 fi
 
+if test "x`smbd -b | grep NSS_WRAPPER`" = "x"; then
+	echo "***"
+	echo "*** You must include --enable-nss-wrapper when compiling Samba"
+	echo "*** in order to execute 'make test'.  Exiting...."
+	echo "***"
+	exit 1
+fi
+
+
 ## 
 ## create the test directory layout
 ##
-echo -n "CREATE TEST ENVIRONMENT IN '$PREFIX'"...
+printf "%s" "CREATE TEST ENVIRONMENT IN '$PREFIX'"...
 /bin/rm -rf $PREFIX/*
-mkdir -p $PRIVATEDIR $LIBDIR $PIDDIR $LOCKDIR $LOGDIR $SOCKET_WRAPPER_DIR
+mkdir -p $PRIVATEDIR $LIBDIR $PIDDIR $LOCKDIR $LOGDIR
+mkdir -p $SOCKET_WRAPPER_DIR
+mkdir -p $WINBINDD_SOCKET_DIR
+chmod 755 $WINBINDD_SOCKET_DIR
 mkdir -p $PREFIX_ABS/tmp
 chmod 777 $PREFIX_ABS/tmp
 
@@ -87,15 +121,13 @@ chmod 777 $PREFIX_ABS/tmp
 ##
 
 cat >$COMMONCONFFILE<<EOF
-	workgroup = SAMBA-TEST
+	workgroup = $WORKGROUP
 
 	private dir = $PRIVATEDIR
 	pid directory = $PIDDIR
 	lock directory = $LOCKDIR
 	log file = $LOGDIR/log.%m
 	log level = 0
-
-	passdb backend = tdbsam
 
 	name resolve order = bcast
 EOF
@@ -108,6 +140,8 @@ cat >$CONFFILE<<EOF
 	interfaces = $TORTURE_INTERFACES
 	panic action = $SCRIPTDIR/gdb_backtrace %d %\$(MAKE_TEST_BINARY)
 	include = $COMMONCONFFILE
+
+	passdb backend = tdbsam
 EOF
 
 cat >$SAMBA4CONFFILE<<EOF
@@ -126,23 +160,38 @@ cat >$SERVERCONFFILE<<EOF
 	panic action = $SCRIPTDIR/gdb_backtrace %d %\$(MAKE_TEST_BINARY)
 	include = $COMMONCONFFILE
 
-	; Necessary to add the build farm hacks
-	add user script = /bin/false
-	add machine script = /bin/false
+	passdb backend = tdbsam
+
+	domain master = yes
+	domain logons = yes
+	time server = yes
+
+	add user script = $PERL $SRCDIR/lib/nss_wrapper/nss_wrapper.pl --path $NSS_WRAPPER_PASSWD --type passwd --action add --name %u
+	add machine script = $PERL $SRCDIR/lib/nss_wrapper/nss_wrapper.pl --path $NSS_WRAPPER_PASSWD --type passwd --action add --name %u
+	delete user script = $PERL $SRCDIR/lib/nss_wrapper/nss_wrapper.pl --path $NSS_WRAPPER_PASSWD --type passwd --action delete --name %u
 
 	kernel oplocks = no
+	kernel change notify = no
 
 	syslog = no
 	printing = bsd
 	printcap name = /dev/null
 
+	winbindd:socket dir = $WINBINDD_SOCKET_DIR
+	idmap uid = 100000-200000
+	idmap gid = 100000-200000
+
+#	min receivefile size = 4000
+
 [tmp]
 	path = $PREFIX_ABS/tmp
 	read only = no
 	smbd:sharedelay = 100000
+	smbd:writetimeupdatedelay = 500000
 	map hidden = yes
 	map system = yes
 	create mask = 755
+	vfs objects = $BINDIR/xattr_tdb.so $BINDIR/streams_xattr.so
 [hideunread]
 	copy = tmp
 	hide unreadable = yes
@@ -165,20 +214,35 @@ EOF
 ## create a test account
 ##
 
+cat >$NSS_WRAPPER_PASSWD<<EOF
+nobody:x:65534:65533:nobody gecos:$PREFIX_ABS:/bin/false
+$USERNAME:x:$USERID:$GROUPID:$USERNAME gecos:$PREFIX_ABS:/bin/false
+EOF
+
+cat >$NSS_WRAPPER_GROUP<<EOF
+nobody:x:65533:
+nogroup:x:65534:nobody
+$USERNAME-group:x:$GROUPID:
+EOF
+
+MAKE_TEST_BINARY="bin/smbpasswd"
+export MAKE_TEST_BINARY
+
 (echo $PASSWORD; echo $PASSWORD) | \
-	smbpasswd -c $CONFFILE -L -s -a $USERNAME >/dev/null || exit 1
+	bin/smbpasswd -c $CONFFILE -L -s -a $USERNAME >/dev/null || exit 1
 
 echo "DONE";
+
+MAKE_TEST_BINARY=""
 
 SERVER_TEST_FIFO="$PREFIX/server_test.fifo"
 export SERVER_TEST_FIFO
 NMBD_TEST_LOG="$PREFIX/nmbd_test.log"
 export NMBD_TEST_LOG
+WINBINDD_TEST_LOG="$PREFIX/winbindd_test.log"
+export WINBINDD_TEST_LOG
 SMBD_TEST_LOG="$PREFIX/smbd_test.log"
 export SMBD_TEST_LOG
-
-MAKE_TEST_BINARY=""
-export MAKE_TEST_BINARY
 
 # start off with 0 failures
 failed=0
@@ -190,6 +254,7 @@ SOCKET_WRAPPER_DEFAULT_IFACE=2
 export SOCKET_WRAPPER_DEFAULT_IFACE
 samba3_check_or_start
 
+
 # ensure any one smbtorture call doesn't run too long
 # and smbtorture will use 127.0.0.6 as source address by default
 SOCKET_WRAPPER_DEFAULT_IFACE=6
@@ -197,6 +262,7 @@ export SOCKET_WRAPPER_DEFAULT_IFACE
 TORTURE4_OPTIONS="$SAMBA4CONFIGURATION"
 TORTURE4_OPTIONS="$TORTURE4_OPTIONS --maximum-runtime=$TORTURE_MAXTIME"
 TORTURE4_OPTIONS="$TORTURE4_OPTIONS --target=samba3"
+TORTURE4_OPTIONS="$TORTURE4_OPTIONS --option=torture:localdir=$PREFIX_ABS/tmp"
 export TORTURE4_OPTIONS
 
 if [ x"$RUN_FROM_BUILD_FARM" = x"yes" ];then
@@ -212,8 +278,9 @@ START=`date`
 (
  # give time for nbt server to register its names
  echo "delaying for nbt name registration"
- sleep 4
+ sleep 10
  # This will return quickly when things are up, but be slow if we need to wait for (eg) SSL init 
+ MAKE_TEST_BINARY="bin/nmblookup"
  bin/nmblookup $CONFIGURATION -U $SERVER_IP __SAMBA__
  bin/nmblookup $CONFIGURATION __SAMBA__
  bin/nmblookup $CONFIGURATION -U 127.255.255.255 __SAMBA__
@@ -221,12 +288,14 @@ START=`date`
  bin/nmblookup $CONFIGURATION $SERVER
  # make sure smbd is also up set
  echo "wait for smbd"
+ MAKE_TEST_BINARY="bin/smbclient"
  bin/smbclient $CONFIGURATION -L $SERVER_IP -U% -p 139 | head -2
  bin/smbclient $CONFIGURATION -L $SERVER_IP -U% -p 139 | head -2
+ MAKE_TEST_BINARY=""
 
  failed=0
 
- . $SCRIPTDIR/tests_$TESTS.sh
+ . $SCRIPTDIR/tests_$SUBTESTS.sh
  exit $failed
 )
 failed=$?
